Extract JS utility functions in separate file
